LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_releasepointercapture_invalid_pointerid-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Event: releasePointerCapture() DOMException - NotFoundError

Test Description: Upon invocation of the releasePointerCapture method, if the provided pointerId value does not match any of the active pointers, a DOMException with the name NotFoundError must be thrown.


Use the mouse, touch or pen to move over or contact this box.
Test complete: Scroll to Summary to view Pass/Fail Results.

The following pointer types were detected: (none).

Refresh the page to run the tests again with a different pointer type.


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN releasePointerCapture: DOMException NotFoundError
